+++ cdx_core/spec/middleware/microscopy_results/presenter_spec.rb
@@ -49,7 +49,7 @@ describe MicroscopyResults::Presenter do
           Extras::Dates::Format.datetime_with_time_zone(MicroscopyResult.first.sample_collected_at, :full_time),
           MicroscopyResult.first.examined_by,
           Extras::Dates::Format.datetime_with_time_zone(MicroscopyResult.first.result_at, :full_time),
-          MicroscopyResult.first.specimen_type.blank? ? "" : I18n.t("test_results.index.specimen_type.#{MicroscopyResult.first.specimen_type}"),
+          MicroscopyResult.first.specimen_type.blank? ? '' : I18n.t("test_results.index.specimen_type.#{MicroscopyResult.first.specimen_type}"),
           MicroscopyResult.first.serial_number,
           Extras::Select.find(MicroscopyResult.test_result_options, MicroscopyResult.first.test_result),
           Extras::Select.find(MicroscopyResult.visual_appearance_options, MicroscopyResult.first.appearance),
