+++ cdx_core/spec/middleware/sites/csv_generator_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-describe CsvGenerator::Sites do
+describe Sites::CsvGenerator do
   let(:user)        { User.make }
   let(:institution) { user.institutions.make }
   let!(:sites)      { 4.times { institution.sites.make } }
@@ -9,8 +9,8 @@ describe CsvGenerator::Sites do
   subject           { described_class.new(Site.all, 'my_custom_name') }
 
   it 'should return a valid CSV with sites' do
-    csv  = CSV.parse(subject.build_csv)
-    expect(csv[0]).to eq(["Name", "Address","City", "State", "Zipcode"])
+    csv = CSV.parse(subject.build_csv)
+    expect(csv[0]).to eq(%w(Name Address City State Zipcode))
     expect(csv[1]).to eq([site.name, site.address, site.city, site.state, site.zip_code])
   end
